moved dag_configuration to airflow_logic
