+++ airflow_logic/dags/hdx_adm_dags.py
@@ -1,7 +1,7 @@
 from airflow import DAG
 
 from dags_utils.utils.config_parser import config
-from dags_utils.utils.dag_configuration import (
+from dags.dags_utils.dag_configuration import (
     get_catchup,
     get_default_arguments,
     get_schedule_interval,
